Clarify Intercom Agent Tool module boundaries
