+++ workers/intercom-escalate-customer-issue/src/create-ticket.ts
@@ -1,64 +1,43 @@
 import { createHash } from "node:crypto"
-import { intercomAppBaseUrl, type RuntimeConfig } from "./config.js"
-import { isDefiniteMutationRejection } from "./http.js"
+import type { RuntimeConfig } from "./config.js"
 import {
   intercomNoteDigest,
+  isDefiniteIntercomMutationRejection,
   normalizeIntercomConversationReference,
   type ConversationSnapshot,
-  type IntercomClient,
-  type IntercomCompany,
-  type IntercomContact,
-  type IntercomIdentity,
-  type IntercomTag,
-  type IntercomTeam,
 } from "./intercom.js"
 import {
   createTicketPage,
   NotionAdapterError,
   NotionCreateError,
   queryTicketsBySourceKey,
-  resolveSyncedConversationPage,
   retrieveAndVerifyTicketPage,
-  retrieveTicketDataSourceSchema,
   type NotionClientLike,
   type TicketDataSourceSchema,
   type TicketPageReference,
 } from "./notion.js"
+import {
+  boundedText,
+  conversationInspectionVersion,
+  hasConfiguredTag,
+  hasNote,
+  inspectDeployment,
+  intercomConversationUrl,
+  lookupDisplayContext,
+  sortedTags,
+  sourceKey,
+  ticketNoteBody,
+  uniqueTicket,
+  type EscalationDependencies,
+} from "./inspect-conversation.js"
 import type {
   CreateTicketInput,
   CreateTicketResult,
-  InspectConversationInput,
-  InspectConversationResult,
   TicketDraft,
 } from "./types.js"
-import { WorkflowError } from "./types.js"
+import { EscalationError } from "./types.js"
 
-const CONTROL_CHARACTER = /[\u0000-\u0008\u000b\u000c\u000e-\u001f\u007f]/
 const INSPECTION_VERSION = /^iv1_[0-9a-f]{64}$/
-
-type IntercomGateway = Pick<
-  IntercomClient,
-  | "getIdentity"
-  | "getTeam"
-  | "getTag"
-  | "getConversation"
-  | "getContact"
-  | "getCompany"
-  | "addTag"
-  | "routeToTeam"
-  | "addInternalNote"
->
-
-export interface WorkflowDependencies {
-  notion: NotionClientLike
-  intercom: IntercomGateway
-}
-
-interface DeploymentContext {
-  schema: TicketDataSourceSchema
-  team: IntercomTeam
-  tag: IntercomTag
-}
 
 interface NormalizedCreateInput {
   conversationId: string
@@ -85,66 +64,11 @@ function sha256(value: string): string {
   return createHash("sha256").update(value).digest("hex")
 }
 
-export function sourceKey(workspaceId: string, conversationId: string): string {
-  return `intercom:${workspaceId}:conversation:${conversationId}`
-}
-
 function sameNotionPageId(left: string, right: string): boolean {
   return (
     left.replaceAll("-", "").toLowerCase() ===
     right.replaceAll("-", "").toLowerCase()
   )
-}
-
-function sortedTags(tags: IntercomTag[]): { id: string; name: string }[] {
-  return tags
-    .map((tag) => ({ id: tag.id, name: tag.name }))
-    .sort((left, right) =>
-      left.id === right.id
-        ? left.name.localeCompare(right.name)
-        : left.id.localeCompare(right.id)
-    )
-}
-
-/**
- * An inspection version is a freshness proof, not an idempotency key. It binds
- * the reviewed conversation to this Worker's fixed destination and route.
- */
-export function conversationInspectionVersion(
-  snapshot: ConversationSnapshot,
-  config: RuntimeConfig,
-  expectedTicketPageId: string | null = null
-): string {
-  return `iv1_${sha256(
-    JSON.stringify({
-      contract: "intercom-notion-ticket-inspection-v1",
-      workspaceId: config.intercomWorkspaceId,
-      conversationId: snapshot.id,
-      destinationDataSourceId: config.notionTicketsDataSourceId,
-      targetTeamId: config.intercomTeamId,
-      targetTagId: config.intercomTagId,
-      expectedTicketPageId: expectedTicketPageId
-        ? expectedTicketPageId.replaceAll("-", "").toLowerCase()
-        : null,
-      createdAt: snapshot.createdAt,
-      updatedAt: snapshot.updatedAt,
-      state: snapshot.state,
-      priority: snapshot.priority,
-      title: snapshot.title,
-      openingMessage: snapshot.openingMessage,
-      contactIds: [...snapshot.contactIds].sort(),
-      companyId: snapshot.companyId,
-      teamAssigneeId: snapshot.teamAssigneeId,
-      slaStatus: snapshot.slaStatus,
-      tags: sortedTags(snapshot.tags),
-      customerEvidence: snapshot.customerEvidence,
-      evidenceTruncated: snapshot.evidenceTruncated,
-      partsTruncated: snapshot.partsTruncated,
-      internalNoteDigests: [...snapshot.internalNoteDigests].sort((a, b) =>
-        a.partId.localeCompare(b.partId)
-      ),
-    })
-  )}`
 }
 
 function continuityVersion(
@@ -176,17 +100,6 @@ function continuityVersion(
   )
 }
 
-function hasConfiguredTag(
-  snapshot: ConversationSnapshot,
-  config: RuntimeConfig
-): boolean {
-  return snapshot.tags.some((tag) => tag.id === config.intercomTagId)
-}
-
-function hasNote(snapshot: ConversationSnapshot, digest: string): boolean {
-  return snapshot.internalNoteDigests.some((note) => note.digest === digest)
-}
-
 function expectedLiveState(
   snapshot: ConversationSnapshot,
   config: RuntimeConfig,
@@ -213,7 +126,7 @@ function assertExpectedLiveState(
     hasConfiguredTag(snapshot, config) !== expected.tagPresent ||
     hasNote(snapshot, desiredNoteDigest) !== expected.notePresent
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "CONVERSATION_CHANGED",
       "The Intercom conversation changed during the action. Inspect it again before deciding whether to continue.",
       "conflict"
@@ -231,32 +144,13 @@ function assertInspectionVersion(
     conversationInspectionVersion(snapshot, config, expectedTicketPageId) !==
     version
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "CONVERSATION_CHANGED",
       "The Intercom conversation changed after inspection. Inspect it again before creating or routing a ticket.",
       "conflict"
     )
   }
 }
-
-function boundedText(value: unknown, label: string, maximum: number): string {
-  if (typeof value !== "string") {
-    throw new WorkflowError("INVALID_INPUT", `${label} must be text.`)
-  }
-  const normalized = value.replace(/\r\n?/g, "\n").trim()
-  if (
-    normalized.length < 1 ||
-    normalized.length > maximum ||
-    CONTROL_CHARACTER.test(normalized)
-  ) {
-    throw new WorkflowError(
-      "INVALID_INPUT",
-      `${label} must contain 1–${maximum} characters of plain text.`
-    )
-  }
-  return normalized
-}
-
 function nullableText(
   value: unknown,
   label: string,
@@ -269,7 +163,7 @@ function nullableText(
 function normalizeTicketDraft(value: TicketDraft | null): TicketDraft | null {
   if (value === null) return null
   if (!value || typeof value !== "object" || Array.isArray(value)) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "INVALID_INPUT",
       "ticketDraft must be a reviewed ticket draft or null."
     )
@@ -282,7 +176,7 @@ function normalizeTicketDraft(value: TicketDraft | null): TicketDraft | null {
       value.priority === "P3"
     )
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "INVALID_INPUT",
       "ticketDraft.priority must be P0, P1, P2, or P3."
     )
@@ -291,7 +185,7 @@ function normalizeTicketDraft(value: TicketDraft | null): TicketDraft | null {
     !Array.isArray(value.reproductionSteps) ||
     value.reproductionSteps.length > 12
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "INVALID_INPUT",
       "ticketDraft.reproductionSteps must contain at most 12 steps."
     )
@@ -320,7 +214,7 @@ function normalizeCreateInput(
     typeof input.inspectionVersion !== "string" ||
     !INSPECTION_VERSION.test(input.inspectionVersion)
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "INVALID_INPUT",
       "inspectionVersion must be the exact opaque value returned by inspection."
     )
@@ -337,211 +231,6 @@ function normalizeCreateInput(
     ticketDraft: normalizeTicketDraft(input.ticketDraft),
   }
 }
-
-function assertConfiguredIdentity(
-  identity: IntercomIdentity,
-  config: RuntimeConfig
-): void {
-  if (
-    identity.workspaceId !== config.intercomWorkspaceId ||
-    identity.adminId !== config.intercomAdminId
-  ) {
-    throw new WorkflowError(
-      "INTERCOM_IDENTITY_MISMATCH",
-      "The Intercom credential does not match the configured workspace and admin."
-    )
-  }
-}
-
-async function inspectDeployment(
-  config: RuntimeConfig,
-  dependencies: WorkflowDependencies
-): Promise<DeploymentContext> {
-  const [identity, team, tag, schema] = await Promise.all([
-    dependencies.intercom.getIdentity(),
-    dependencies.intercom.getTeam(config.intercomTeamId),
-    dependencies.intercom.getTag(config.intercomTagId),
-    retrieveTicketDataSourceSchema(
-      dependencies.notion,
-      config.notionTicketsDataSourceId
-    ),
-  ])
-  assertConfiguredIdentity(identity, config)
-  if (team.id !== config.intercomTeamId || tag.id !== config.intercomTagId) {
-    throw new WorkflowError(
-      "INTERCOM_ROUTE_MISMATCH",
-      "Intercom returned a different configured team or tag."
-    )
-  }
-  return { schema, team, tag }
-}
-
-function intercomConversationUrl(
-  region: RuntimeConfig["intercomRegion"],
-  workspaceId: string,
-  conversationId: string
-): string {
-  return `${intercomAppBaseUrl(region)}/a/inbox/${encodeURIComponent(
-    workspaceId
-  )}/inbox/shared/all/conversation/${encodeURIComponent(conversationId)}`
-}
-
-function normalizeInspectionReference(
-  input: InspectConversationInput,
-  config: RuntimeConfig
-): { conversationPageId: string | null; conversationId: string | null } {
-  const conversationPageId =
-    input.conversationPageId === null
-      ? null
-      : boundedText(input.conversationPageId, "conversationPageId", 2_000)
-  const conversationId =
-    input.conversationId === null
-      ? null
-      : normalizeIntercomConversationReference(input.conversationId, {
-          region: config.intercomRegion,
-          workspaceId: config.intercomWorkspaceId,
-        })
-  if ((conversationPageId === null) === (conversationId === null)) {
-    throw new WorkflowError(
-      "INVALID_INPUT",
-      "Provide exactly one of conversationPageId or conversationId."
-    )
-  }
-  return { conversationPageId, conversationId }
-}
-
-async function lookupDisplayContext(
-  intercom: IntercomGateway,
-  snapshot: ConversationSnapshot
-): Promise<{
-  customer: IntercomContact | null
-  company: IntercomCompany | null
-}> {
-  const [customer, company] = await Promise.all([
-    snapshot.contactIds[0]
-      ? intercom.getContact(snapshot.contactIds[0])
-      : Promise.resolve(null),
-    snapshot.companyId
-      ? intercom.getCompany(snapshot.companyId)
-      : Promise.resolve(null),
-  ])
-  return { customer, company }
-}
-
-function uniqueTicket(
-  tickets: TicketPageReference[]
-): TicketPageReference | null {
-  if (tickets.length > 1) {
-    throw new WorkflowError(
-      "DUPLICATE_NOTION_TICKETS",
-      "More than one Notion ticket has the same Intercom source key. Resolve the duplicates before continuing.",
-      "conflict"
-    )
-  }
-  return tickets[0] ?? null
-}
-
-function noteMarker(key: string, pageId: string): string {
-  const canonicalPageId = pageId.replaceAll("-", "").toLowerCase()
-  return `icn_${sha256(`${key}:${canonicalPageId}`).slice(0, 32)}`
-}
-
-export function ticketNoteBody(key: string, pageId: string): string {
-  const canonicalPageId = pageId.replaceAll("-", "").toLowerCase()
-  return `Notion ticket: https://www.notion.so/${canonicalPageId}\nReference: ${noteMarker(
-    key,
-    pageId
-  )}`
-}
-
-export async function inspectIntercomConversation(
-  input: InspectConversationInput,
-  config: RuntimeConfig,
-  dependencies: WorkflowDependencies
-): Promise<InspectConversationResult> {
-  const reference = normalizeInspectionReference(input, config)
-  const [sourcePage, deployment] = await Promise.all([
-    reference.conversationPageId
-      ? resolveSyncedConversationPage(
-          dependencies.notion,
-          reference.conversationPageId
-        )
-      : Promise.resolve(null),
-    inspectDeployment(config, dependencies),
-  ])
-  const conversationId = normalizeIntercomConversationReference(
-    sourcePage?.conversationId ?? (reference.conversationId as string),
-    {
-      region: config.intercomRegion,
-      workspaceId: config.intercomWorkspaceId,
-    }
-  )
-  const snapshot = await dependencies.intercom.getConversation(conversationId)
-  const key = sourceKey(config.intercomWorkspaceId, conversationId)
-  const [display, tickets] = await Promise.all([
-    lookupDisplayContext(dependencies.intercom, snapshot),
-    queryTicketsBySourceKey(dependencies.notion, deployment.schema, key),
-  ])
-  const existing = uniqueTicket(tickets)
-  const exactNotePresent = existing
-    ? hasNote(
-        snapshot,
-        intercomNoteDigest(ticketNoteBody(key, existing.pageId))
-      )
-    : false
-  const routeComplete = Boolean(
-    existing &&
-      snapshot.teamAssigneeId === config.intercomTeamId &&
-      hasConfiguredTag(snapshot, config) &&
-      exactNotePresent
-  )
-
-  return {
-    conversationId,
-    intercomUrl: intercomConversationUrl(
-      config.intercomRegion,
-      config.intercomWorkspaceId,
-      conversationId
-    ),
-    sourcePageId: sourcePage?.pageId ?? null,
-    sourcePageUrl: sourcePage?.pageUrl ?? null,
-    inspectionVersion: conversationInspectionVersion(
-      snapshot,
-      config,
-      existing?.pageId ?? null
-    ),
-    state: snapshot.state,
-    priority: snapshot.priority,
-    title: snapshot.title,
-    openingMessage: snapshot.openingMessage,
-    customer: display.customer,
-    company: display.company,
-    currentTeamId: snapshot.teamAssigneeId,
-    slaStatus: snapshot.slaStatus,
-    tags: snapshot.tags,
-    evidence: snapshot.customerEvidence,
-    evidenceTruncated: snapshot.evidenceTruncated,
-    partsTruncated: snapshot.partsTruncated,
-    existingTicket: existing
-      ? { pageId: existing.pageId, url: existing.pageUrl }
-      : null,
-    ticketCreationState: existing ? "existing" : "none",
-    plannedRoute: {
-      teamId: deployment.team.id,
-      teamName: deployment.team.name,
-      tagId: deployment.tag.id,
-      tagName: deployment.tag.name,
-    },
-    message: existing
-      ? routeComplete
-        ? "The Notion ticket and configured internal route are already complete. Show the existing ticket link."
-        : snapshot.partsTruncated && !exactNotePresent
-          ? "A Notion ticket exists, but the bounded Intercom history cannot prove whether its internal link note exists. Show the ticket and ask the user to verify the note in Intercom before attempting repair."
-          : "A Notion ticket already exists. Show its link and, if the user confirms, reuse it to finish the configured internal route without overwriting the ticket."
-      : "Draft and show the ticket to the user. Call createNotionTicket only after the user confirms it.",
-  }
-}
-
 function changed(progress: CreateProgress): boolean | null {
   if (progress.knownChanged) return true
   return progress.uncertainWrite ? null : false
@@ -570,11 +259,11 @@ function completedResult(progress: CreateProgress): CreateTicketResult {
 
 function isExpectedError(
   error: unknown
-): error is WorkflowError | NotionAdapterError {
-  return error instanceof WorkflowError || error instanceof NotionAdapterError
+): error is EscalationError | NotionAdapterError {
+  return error instanceof EscalationError || error instanceof NotionAdapterError
 }
 
-function nextStepFor(error: WorkflowError | NotionAdapterError): string {
+function nextStepFor(error: EscalationError | NotionAdapterError): string {
   if (error.code === "NOTION_CREATE_OUTCOME_UNKNOWN") {
     return "Search the configured Notion data source for this conversation's exact Intercom source key. Do not issue another create until the first outcome is resolved."
   }
@@ -613,11 +302,11 @@ function nextStepFor(error: WorkflowError | NotionAdapterError): string {
 }
 
 function failureResult(
-  error: WorkflowError | NotionAdapterError,
+  error: EscalationError | NotionAdapterError,
   progress: CreateProgress
 ): CreateTicketResult {
   let status: CreateTicketResult["status"] =
-    error instanceof WorkflowError ? error.status : "blocked"
+    error instanceof EscalationError ? error.status : "blocked"
   if (error.code === "NOTION_QUERY_NOT_UNIQUE") status = "conflict"
   if (progress.ticket.pageId && status === "blocked") status = "partial_failure"
   const exactRetryIsSafe =
@@ -637,7 +326,7 @@ function failureResult(
 }
 
 async function readAfterIntercomMutation(
-  dependencies: WorkflowDependencies,
+  dependencies: EscalationDependencies,
   conversationId: string,
   progress: CreateProgress,
   action: keyof CreateTicketResult["intercom"]
@@ -647,7 +336,7 @@ async function readAfterIntercomMutation(
   } catch (error) {
     if (!isExpectedError(error)) throw error
     progress.intercom[action] = "unknown"
-    throw new WorkflowError(
+    throw new EscalationError(
       action === "note"
         ? "INTERCOM_NOTE_OUTCOME_UNKNOWN"
         : "INTERCOM_MUTATION_OUTCOME_UNKNOWN",
@@ -671,7 +360,7 @@ async function reconcileUnknownNotionCreate(
   try {
     tickets = await queryTicketsBySourceKey(notion, schema, key)
   } catch {
-    throw new WorkflowError(
+    throw new EscalationError(
       "NOTION_CREATE_OUTCOME_UNKNOWN",
       "Notion may have created the ticket, but the exact source-key lookup could not prove the outcome.",
       "ambiguous",
@@ -685,14 +374,14 @@ async function reconcileUnknownNotionCreate(
     expectedPageId &&
     !sameNotionPageId(ticket.pageId, expectedPageId)
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "NOTION_TICKET_MISMATCH",
       "Notion source-key lookup returned a different ticket than the page identified by the create response.",
       "conflict"
     )
   }
   if (!ticket) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "NOTION_CREATE_OUTCOME_UNKNOWN",
       "Notion may have created the ticket, but no exact source-key match is visible yet.",
       "ambiguous",
@@ -713,7 +402,7 @@ async function confirmSoleCreatedTicket(
   try {
     tickets = await queryTicketsBySourceKey(notion, schema, key)
   } catch {
-    throw new WorkflowError(
+    throw new EscalationError(
       "NOTION_CREATE_OUTCOME_UNKNOWN",
       "Notion created the ticket, but the Worker could not prove that it is the sole exact source-key match.",
       "ambiguous",
@@ -723,7 +412,7 @@ async function confirmSoleCreatedTicket(
   }
   const unique = uniqueTicket(tickets)
   if (!unique || !sameNotionPageId(unique.pageId, created.pageId)) {
-    throw new WorkflowError(
+    throw new EscalationError(
       unique ? "NOTION_TICKET_MISMATCH" : "NOTION_CREATE_OUTCOME_UNKNOWN",
       unique
         ? "Notion source-key lookup returned a different ticket after creation."
@@ -743,7 +432,7 @@ async function assertSoleTicket(
 ): Promise<void> {
   const live = uniqueTicket(await queryTicketsBySourceKey(notion, schema, key))
   if (!live || !sameNotionPageId(live.pageId, expected.pageId)) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "NOTION_TICKET_MISMATCH",
       "The expected Notion ticket is not the sole exact source-key match.",
       "conflict"
@@ -756,7 +445,7 @@ async function ensureTag(
   expected: ExpectedLiveState,
   desiredNoteDigest: string,
   config: RuntimeConfig,
-  dependencies: WorkflowDependencies,
+  dependencies: EscalationDependencies,
   progress: CreateProgress
 ): Promise<{ snapshot: ConversationSnapshot; expected: ExpectedLiveState }> {
   assertExpectedLiveState(snapshot, expected, config, desiredNoteDigest)
@@ -771,10 +460,10 @@ async function ensureTag(
     responseConfirmed = true
     progress.knownChanged = true
   } catch (error) {
-    if (!isExpectedError(error) || isDefiniteMutationRejection(error)) {
+    if (!isExpectedError(error) || isDefiniteIntercomMutationRejection(error)) {
       throw error
     }
-    if (!(error instanceof WorkflowError) || !error.ambiguous) throw error
+    if (!(error instanceof EscalationError) || !error.ambiguous) throw error
     progress.uncertainWrite = true
   }
 
@@ -791,7 +480,7 @@ async function ensureTag(
   }
   if (!hasConfiguredTag(refreshed, config)) {
     progress.intercom.tag = "unknown"
-    throw new WorkflowError(
+    throw new EscalationError(
       "INTERCOM_MUTATION_OUTCOME_UNKNOWN",
       "Intercom did not prove that the configured escalation tag is present.",
       "ambiguous",
@@ -810,7 +499,7 @@ async function ensureRoute(
   expected: ExpectedLiveState,
   desiredNoteDigest: string,
   config: RuntimeConfig,
-  dependencies: WorkflowDependencies,
+  dependencies: EscalationDependencies,
   progress: CreateProgress
 ): Promise<{ snapshot: ConversationSnapshot; expected: ExpectedLiveState }> {
   assertExpectedLiveState(snapshot, expected, config, desiredNoteDigest)
@@ -825,10 +514,10 @@ async function ensureRoute(
     responseConfirmed = true
     progress.knownChanged = true
   } catch (error) {
-    if (!isExpectedError(error) || isDefiniteMutationRejection(error)) {
+    if (!isExpectedError(error) || isDefiniteIntercomMutationRejection(error)) {
       throw error
     }
-    if (!(error instanceof WorkflowError) || !error.ambiguous) throw error
+    if (!(error instanceof EscalationError) || !error.ambiguous) throw error
     progress.uncertainWrite = true
   }
 
@@ -845,7 +534,7 @@ async function ensureRoute(
   }
   if (refreshed.teamAssigneeId !== config.intercomTeamId) {
     progress.intercom.route = "unknown"
-    throw new WorkflowError(
+    throw new EscalationError(
       "INTERCOM_MUTATION_OUTCOME_UNKNOWN",
       "Intercom did not prove that the conversation reached the configured team.",
       "ambiguous",
@@ -866,7 +555,7 @@ async function ensureNote(
   desiredNoteDigest: string,
   allowFirstNoteWhenHistoryIsTruncated: boolean,
   config: RuntimeConfig,
-  dependencies: WorkflowDependencies,
+  dependencies: EscalationDependencies,
   progress: CreateProgress
 ): Promise<{ snapshot: ConversationSnapshot; expected: ExpectedLiveState }> {
   assertExpectedLiveState(snapshot, expected, config, desiredNoteDigest)
@@ -876,7 +565,7 @@ async function ensureNote(
   }
   if (snapshot.partsTruncated && !allowFirstNoteWhenHistoryIsTruncated) {
     progress.intercom.note = "unknown"
-    throw new WorkflowError(
+    throw new EscalationError(
       "INTERCOM_NOTE_NOT_VISIBLE",
       "Intercom omitted older conversation parts, so the Worker cannot prove that the exact ticket-link note is absent.",
       "ambiguous",
@@ -891,10 +580,10 @@ async function ensureNote(
     responseConfirmed = true
     progress.knownChanged = true
   } catch (error) {
-    if (!isExpectedError(error) || isDefiniteMutationRejection(error)) {
+    if (!isExpectedError(error) || isDefiniteIntercomMutationRejection(error)) {
       throw error
     }
-    if (!(error instanceof WorkflowError) || !error.ambiguous) throw error
+    if (!(error instanceof EscalationError) || !error.ambiguous) throw error
     progress.uncertainWrite = true
   }
 
@@ -910,7 +599,7 @@ async function ensureNote(
   }
   if (!hasNote(refreshed, desiredNoteDigest)) {
     progress.intercom.note = "unknown"
-    throw new WorkflowError(
+    throw new EscalationError(
       "INTERCOM_NOTE_OUTCOME_UNKNOWN",
       "Intercom may have added the internal ticket-link note, but the exact note is not visible.",
       "ambiguous",
@@ -927,7 +616,7 @@ async function ensureNote(
 async function createNotionTicketCore(
   input: NormalizedCreateInput,
   config: RuntimeConfig,
-  dependencies: WorkflowDependencies,
+  dependencies: EscalationDependencies,
   progress: CreateProgress
 ): Promise<CreateTicketResult> {
   const deployment = await inspectDeployment(config, dependencies)
@@ -953,7 +642,7 @@ async function createNotionTicketCore(
     }
   } else {
     if (!input.ticketDraft) {
-      throw new WorkflowError(
+      throw new EscalationError(
         "TICKET_DRAFT_REQUIRED",
         "No Notion ticket exists, so a reviewed ticketDraft is required before creation."
       )
@@ -976,7 +665,7 @@ async function createNotionTicketCore(
         )
       )
     ) {
-      throw new WorkflowError(
+      throw new EscalationError(
         "NOTION_TICKET_STATE_CHANGED",
         "A Notion ticket appeared after inspection. Inspect the conversation again and review that ticket before routing it.",
         "conflict"
@@ -1084,7 +773,7 @@ async function createNotionTicketCore(
     progress.ticket.action !== "created"
   ) {
     progress.intercom.note = "unknown"
-    throw new WorkflowError(
+    throw new EscalationError(
       "INTERCOM_NOTE_NOT_VISIBLE",
       "Intercom omitted older conversation parts, so the Worker cannot prove that the exact ticket-link note is absent.",
       "ambiguous",
@@ -1136,7 +825,7 @@ async function createNotionTicketCore(
     final.teamAssigneeId !== config.intercomTeamId ||
     !hasNote(final, desiredNoteDigest)
   ) {
-    throw new WorkflowError(
+    throw new EscalationError(
       "INTERCOM_POSTCONDITION_FAILED",
       "Intercom did not preserve every configured postcondition through the final read.",
       "ambiguous",
@@ -1152,7 +841,7 @@ async function createNotionTicketCore(
 export async function createNotionTicket(
   rawInput: CreateTicketInput,
   config: RuntimeConfig,
-  dependencies: WorkflowDependencies
+  dependencies: EscalationDependencies
 ): Promise<CreateTicketResult> {
   const progress: CreateProgress = {
     conversationId:
